pyproject.toml
[tool.poetry]
name = "good-habits-tracker-3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "good_habits_tracker_3"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.8"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
pillow = "^10.1.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = "^5.3.6"
eventlet = "^0.33.3"
django-celery-beat = "^2.5.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
django-cors-headers = "^4.3.1"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
